+++ skeleton-sp17-master/hw1/synthesizer/BoundedQueue.java
@@ -2,12 +2,12 @@ package synthesizer;
 import java.util.Iterator;
 
 
-public interface BoundedQueue extends Iterable{
+public interface  BoundedQueue extends Iterable {
     int capcitity(); //Return size of the buffer.
     int fillCount(); //return number of items currently in the buffer.
-    void enqueue(int x); // add item x to the end.
-    int dequeue(); // delete and return item from the front.
-    int peek(); // Return (but don't delete) item from the front.
+    void enqueue(double x); // add item x to the end.
+    double dequeue(); // delete and return item from the front.
+    double peek(); // Return (but don't delete) item from the front.
     
     // is the buffer empty?
     default boolean isEmpty() {
